backend/pyproject.toml
[project]
name = "rag-agent-platform"
version = "0.1.0"
description = "RAG Agent Platform - Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",
    # Validation & Settings
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    # HTTP Client
    "httpx>=0.28.0",
    # Payment
    "stripe>=11.0.0",
    # Redis
    "redis>=5.0.0",
    # Rate Limiting
    "slowapi>=0.1.9",
    # Utils
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    # Document Processing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    # Storage
    "minio>=7.2.0",
    # Embeddings (using LiteLLM API - no local model needed)
    # OpenTelemetry - Tracing
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    # OpenTelemetry - Logging
    "opentelemetry-instrumentation-logging>=0.48b0",
    # OpenTelemetry - Redis
    "opentelemetry-instrumentation-redis>=0.48b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # Do not perform function call in argument defaults (FastAPI Depends)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
